+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ### Added
 
 - add JetBrains annotations dependency
+- introduced PathfindingSearch which represents a single search operation
 
 ### Changed
 
 - marked getEnvironmentContext NonNull in SearchContext
+- replaced CompletionStage<PathfinderResult> return from Pathfinder with PathfindingSearch
 
 ### Fixed